refactor: split server.ts into app.ts and server.ts
