+++ server.ts
@@ -1,39 +1,7 @@
-import express,  { Request, Response} from 'express';
-import cors from 'cors';
-import cookieParser from 'cookie-parser';
+import { app } from './app';
 import 'dotenv/config';
 import connectDB from './utils/db';
-import todoRouter from './routes/route';
 
-
-
-const app = express();
-
-
-//body parser
-app.use(express.json({limit: "50mb"}));
-
-//cookie parser
-app.use(cookieParser());
-
-//cors
-app.use(cors({
-  origin: process.env.ORIGIN
-}));
-
-// routes
-app.use(
-  "/api/v1",
-  todoRouter,
-);
-
-//testing api
-app.use('/test', (req: Request, res:Response) => {
-  res.status(200).json({
-    success: true,
-    message: "API is working"
-  });
-}); 
 
 // connect server to localhost
 app.listen(process.env.PORT, () => {
